requirements.txt
mpmath==1.1.0
mypy==0.650
mypy-extensions==0.4.1
Pillow==5.3.0
sympy==1.3
typed-ast==1.1.1
